+++ .github/workflows/pr_tests_gpu.yml
@@ -16,6 +16,7 @@ on:
       - "src/diffusers/loaders/peft.py"
       - "tests/pipelines/testing_utils/**.py"
       - "tests/models/test_modeling_common.py"
+      - "tests/models/testing_utils/**.py"
       - "examples/**/*.py"
   workflow_dispatch:
