+++ tests/pipelines/testing_utils/common.py
@@ -19,6 +19,7 @@ import json
 import os
 from typing import Callable
 
+import numpy as np
 import pytest
 import torch
 import torch.nn as nn
@@ -77,6 +78,16 @@ def cast_pipeline_to_dtype(pipe, dtype):
         if isinstance(component, torch.nn.Module):
             cast_module_to_dtype(component, dtype)
     return pipe
+
+
+# Some models (e.g. unCLIP) are extremely likely to significantly deviate depending on which hardware is used.
+# This helper function is used to check that the image doesn't deviate on average more than 10 pixels from a
+# reference image.
+def assert_mean_pixel_difference(image, expected_image, expected_max_diff=10):
+    image = np.asarray(DiffusionPipeline.numpy_to_pil(image)[0], dtype=np.float32)
+    expected_image = np.asarray(DiffusionPipeline.numpy_to_pil(expected_image)[0], dtype=np.float32)
+    avg_diff = np.abs(image - expected_image).mean()
+    assert avg_diff < expected_max_diff, f"Error image deviates {avg_diff} pixels on average"
 
 
 class BasePipelineTesterConfig:
